Background threading implemented

+++ app/src/main/java/com/example/youssefiibrahim/gallereo/model/Pair.java
@@ -2,10 +2,10 @@ package com.example.youssefiibrahim.gallereo.model;
 
 public class Pair {
 
-    public Integer weight;
+    public Double weight;
     public String word;
 
-    public Pair(String word, Integer weight) {
+    public Pair(String word, Double weight) {
         this.word = word;
         this.weight = weight;
     }
